// src/configs/global.enum.ts
export enum GetUsersParcel {
  GET_PARCEL_CALLER = 'GET_PARCEL_CALLER',
  GET_PARCEL_DEFAULT = 'GET_PARCEL_DEFAULT',
  GET_PARCEL_STARTED = 'GET_PARCEL_STARTED',
  GET_PARCEL_FAILED = 'GET_PARCEL_FAILED',
  GET_PARCEL_SUCCESS = 'GET_PARCEL_SUCCESS',
}

export enum verificationActionType {
  SEND_VERIFICATION_CODE_CALLER = 'DO-SEND-VERIFICATION-CODE ',
  SEND_VERIFICATION_CODE_DEFAULT = 'SEND-VERIFICATION-CODE-NOT-STARTED',
  SEND_VERIFICATION_CODE_STARTED = 'SEND-VERIFICATION-CODE-STARTED',
  SEND_VERIFICATION_CODE_FAILED = 'SEND-VERIFICATION-CODE-FAILED',
  SEND_VERIFICATION_CODE_SUCCESS = 'SEND-VERIFICATION-CODE-SUCCESS',
}

export enum createParcel {
  CREATE_PARCEL_CALLER = 'DO-CREATE_PARCEL ',
  CREATE_PARCEL_DEFAULT = 'CREATE_PARCEL-NOT-STARTED',
  CREATE_PARCEL_STARTED = 'CREATE_PARCEL-STARTED',
  CREATE_PARCEL_FAILED = 'CREATE_PARCEL-FAILED',
  CREATE_PARCEL_SUCCESS = 'CREATE_PARCEL-SUCCESS',
}

export enum cancelDeliveryActionType {
  CANCEL_DELIVERY_CALLER = 'DO-CANCEL-DELIVERY ',
  CANCEL_DELIVERY_DEFAULT = 'CANCEL-DELIVERY-NOT-STARTED',
  CANCEL_DELIVERY_STARTED = 'CANCEL-DELIVERY-STARTED',
  CANCEL_DELIVERY_FAILED = 'CANCEL-DELIVERY-FAILED',
  CANCEL_DELIVERY_SUCCESS = 'CANCEL-DELIVERY-SUCCESS',
}

export enum confirmDeliveryActionType {
  CONFIRM_DELIVERY_CALLER = 'DO-CONFIRM_DELIVERY ',
  CONFIRM_DELIVERY_DEFAULT = 'CONFIRM-DELIVERY-NOT-STARTED',
  CONFIRM_DELIVERY_STARTED = 'CONFIRM-DELIVERY-STARTED',
  CONFIRM_DELIVERY_FAILED = 'CONFIRM-DELIVERY-FAILED',
  CONFIRM_DELIVERY_SUCCESS = 'CREATE-PARCEL-SUCCESS',
}

export enum RessetPasswordActionType {
  RESSET_PASSWORD_CALLER = 'DO-RESSET-PASSWORD ',
  RESSET_PASSWORD_DEFAULT = 'RESSET-PASSWORD-NOT-STARTED',
  RESSET_PASSWORD_STARTED = 'RESSET-PASSWORD-STARTED',
  RESSET_PASSWORD_FAILED = 'RESSET-PASSWORD-FAILED',
  RESSET_PASSWORD_SUCCESS = 'RESSET-PASSWORD-SUCCESS',
}

export enum searchLocation {
  SEARCH_LOCATION_CALLER = 'DO-SEARCH-LOCATION ',
  SEARCH_LOCATION_DEFAULT = 'SEARCH-LOCATION-NOT-STARTED',
  SEARCH_LOCATION_STARTED = 'SEARCH-LOCATION-STARTED',
  SEARCH_LOCATION_FAILED = 'SEARCH-LOCATION-FAILED',
  SEARCH_LOCATION_SUCCESS = 'SEARCH-LOCATION-SUCCESS',
}

export enum setLocation {
  SET_LOCATION_CALLER = 'DO-SET_LOCATION ',
  SET_LOCATION_DEFAULT = 'SET_LOCATION-NOT-STARTED',
  SET_LOCATION_STARTED = 'SET_LOCATION-STARTED',
  SET_LOCATION_FAILED = 'SET_LOCATION-FAILED',
  SET_LOCATION_SUCCESS = 'SET_LOCATION-SUCCESS',
}
// ...
export enum currentLocation {
  GET_CUURENT_LOCATION_CALLER = 'DO-GET-CURRENT-LOCATION',
  GET_CUURENT_LOCATION_STARTED = 'GET-CURRENT-LOCATION-STARTED',
  GET_CUURENT_LOCATION_SUCCESS = 'GET-CURRENT-LOCATION-SUCCESS',
  GET_CUURENT_LOCATION_FAILED = 'GET-CURRENT-LOCATION-FAILED',
  GET_CUURENT_LOCATION_DEFAULT = 'GET-CURRENT-LOCATION-NOT-STARTED',
}

export enum inputActionType {
  //SAGA CALLERS
  SET_LOGIN_PHONE_NUMBER_CALLER = 'SET-LOGIN-PHONE-NUMBER',
  SET_LOGIN_PASSWORD_CALLER = 'SET-LOGIN-PASSWORD',
  SET_FIRST_NAME_CALLER = 'SET-FIRST-NAME',
  SET_LAST_NAME_CALLER = 'SET-LAST-NAME',
  SET_SIGNUP_PHONE_NUMBER_CALLER = 'SET-SIGNUP-PHONE-NUMBER',
  SET_SIGNUP_PASSWORD_CALLER = 'SET-SIGNUP-PASSWORD',
  SET_GENDER_CALLER = 'SET-GENDER',
  SET_RESSET_PASSWORD_PHONE_NUMBER_CALLER = 'SET-RESSET-PASSWORD-PHONE-NUMBER',
  SET_VERIFICATION_PIN_CALLER = 'SET-VERIFICATION-PIN',
  SET_TIMER_ELAPSE_CALLER = 'SET-TIMER-ELAPSE',
  CHANGE_COUNTDOWN_COMPONENT_ID_CALLER = 'CHANGE-COUTDOWN-COMPOENT-ID',
  SET_NEW_PASSWORD_CALLER = 'SET-NEW-PASSWORD',
  SET_CONFIRM_NEW_PASSWORD_CALLER = 'SET-CONFIRM-NEW-PASSWORD',
  SET_PARCEL_LOCATION_CALLER = 'DO-SET-PARCEL-LOCATION',
  SET_PARCEL_DESITINATION_CALLER = 'DO-SET-PARCEL-DESTINATION',
  SET_LOCATION_ACTIVE_CALLER = 'DO-SET-LOCATION-ACTIVE',
  SET_VARIABLES_CALLER = 'DO-SET-VARIABLES',
  SET_RESET_INPUTS_CALLER = 'DO-RESET-INPUTS',
  SET_DURATION_AND_DISTANCE_CALLER = 'DO-SET-DURATION-AND-DISTANCE',
  SET_ENABLE_ADD_BUTTON_CALLER = 'DO-ENABLE-ADD-BUTTON',
  SET_ADD_VALID_CARD_CALLER = 'DO-ADD-VALID-CARD',
  SET_REMOVE_VALID_CARD_CALLER = 'DO-REMOVE-VALID-CARD',
  SET_CHANGE_PAYMENT_METHOD_CALLER = 'DO-CHANGE-PAYMENT-METHOD',
  SET_TRACK_ACTIVE_DELIVERY_CALLER = 'DO-SET-DELIVERY-ON-PROGRESS',
  SET_USER_CURRENT_LOCATION_CALLER = 'DO-SET-CURRENT-LOCATION',
  SET_LOCATION_CALLER = 'DO-SET-FETCHED-LOCATION',
  SET_PROFILE_PICTURE_CALLER = 'DO-SET-USER-PROFILE-PICTURE',
  SET_CARD_EMAIL_CALLER = 'DO_SET_CARD_EMAIL_CALLER',

  //ACTIONS
  SET_LOGIN_PHONE_NUMBER = 'SET-LOGIN-PHONE-NUMBER',
  SET_LOGIN_PASSWORD = 'SET-LOGIN-PASSWORD',
  SET_FIRST_NAME = 'SET-FIRST-NAME',
  SET_LAST_NAME = 'SET-LAST-NAME',
  SET_SIGNUP_PHONE_NUMBER = 'SET-SIGNUP-PHONE-NUMBER',
  SET_SIGNUP_PASSWORD = 'SET-SIGNUP-PASSWORD',
  SET_GENDER = 'SET-GENDER',
  SET_RESSET_PASSWORD_PHONE_NUMBER = 'SET-RESSET-PASSWORD-PHONE-NUMBER',
  SET_VERIFICATION_PIN = 'SET-VERIFICATION-PIN',
  SET_TIMER_ELAPSE = 'SET-TIMER-ELAPSE',
  CHANGE_COUNTDOWN_COMPONENT_ID = 'CHANGE-COUTDOWN-COMPOENT-ID',
  SET_NEW_PASSWORD = 'SET-NEW-PASSWORD',
  SET_CONFIRM_NEW_PASSWORD = 'SET-CONFIRM-NEW-PASSWORD',
  SET_PARCEL_LOCATION = 'SET-PARCEL-LOCATION',
  SET_PARCEL_DESITINATION = 'SET-PARCEL-DESTINATION',
  SET_LOCATION_ACTIVE = 'SET-LOCATION-ACTIVE',
  SET_VARIABLES = 'SET-VARIABLES',
  SET_RESET_INPUTS = 'RESET-INPUTS',
  SET_DURATION_AND_DISTANCE = 'SET-DURATION-AND-DISTANCE',
  SET_ENABLE_ADD_BUTTON = 'ENABLE-ADD-BUTTON',
  SET_ADD_VALID_CARD = 'ADD-VALID-CARD',
  SET_REMOVE_VALID_CARD = 'REMOVE-VALID-CARD',
  SET_CHANGE_PAYMENT_METHOD = 'CHANGE-PAYMENT-METHOD',
  SET_TRACK_ACTIVE_DELIVERY = 'SET-DELIVERY-ON-PROGRESS',
  SET_USER_CURRENT_LOCATION = 'SET-CURRENT-LOCATION',
  SET_PROFILE_PICTURE = 'SET-USER-PROFILE-PICTURE',
  SET_CARD_EMAIL = 'SET_CARD_EMAIL',
}

export enum authAction {
  LOGOUT = 'LOGOUT',
}

export enum AddCreditCard {
  ADD_CREDIT_CARD_DEFAULT = 'ADD_CREDEIT_CARD_DEFAULT',
  ADD_CREDEIT_CARD_STARTED = 'ADD_CREDIT_CARD_STARTED',
  ADD_CREDIT_CARD_FAILED = 'ADD_CREDIT_CARD_FAILED',
  ADD_CREDIT_CARD_SUCCESS = 'ADD_CREDIT_CARD_SUCCESS',
  ADD_CREDIT_CARD_CALLER = 'ADD_CREDIT_CARD_CALLER',
}

export enum RemoveCreditCard {
  REMOVE_CREDIT_CARD_DEFAULT = 'REMOVE_CREDEIT_CARD_DEFAULT',
  REMOVE_CREDEIT_CARD_STARTED = 'REMOVE_CREDIT_CARD_STARTED',
  REMOVE_CREDIT_CARD_FAILED = 'REMOVE_CREDIT_CARD_FAILED',
  REMOVE_CREDIT_CARD_SUCCESS = 'REMOVE_CREDIT_CARD_SUCCESS',
  REMOVE_CREDIT_CARD_CALLER = 'REMOVE_CREDIT_CARD_CALLER',
}

export enum GetCreditCards {
  GET_CREDIT_CARD_DEFAULT = 'GET_CREDEIT_CARD_DEFAULT',
  GET_CREDEIT_CARD_STARTED = 'GET_CREDIT_CARD_STARTED',
  GET_CREDIT_CARD_FAILED = 'GET_CREDIT_CARD_FAILED',
  GET_CREDIT_CARD_SUCCESS = 'GET_CREDIT_CARD_SUCCESS',
  GET_CREDIT_CARD_CALLER = 'GET_CREDIT_CARD_CALLER',
}